requirements.txt
fastapi
uvicorn
motor
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
cryptography
email-validator
